backend/requirements.txt
asgiref==3.4.1
Django==4.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==1.2.2
djangorestframework==3.13.1
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
